services/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis==5.0.1
asyncpg==0.29.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
ruff==0.1.13
black==23.12.1
sqlalchemy==2.0.25
alembic==1.13.1
greenlet==3.0.3
passlib[argon2]==1.7.4

python-jose[cryptography]==3.3.0
python-multipart==0.0.6



